+++ .github/workflows/crates-io.yml
@@ -1,4 +1,4 @@
-name: Publish prusti-contracts to Crates.io
+name: Publish on Crates.io
 
 on:
   workflow_dispatch:
