--- .github/workflows/test-on-crates.yml
@@ -1,26 +1,0 @@
-name: Test on crates
-
-on:
-  workflow_dispatch:
-  push:
-    branches: [master, staging, trying]
-    paths-ignore: 'docs/**'
-
-jobs:
-  # Test `cargo-prusti` on a collection of crates.
-  test-crates:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Check out the repo
-        uses: actions/checkout@v2
-      - uses: actions/setup-java@v1
-        with:
-          java-version: '15'
-      - name: Set up the environment
-        run: python x.py setup
-      - name: Cache cargo
-        uses: Swatinem/rust-cache@v1.3.0
-      - name: Build with cargo --release
-        run: python x.py build --release --all --verbose
-      - name: Test Prusti on a collection of crates
-        run: ./target/release/test-crates
