--- prusti-contracts/prusti-contracts-internal/Prusti.toml
@@ -1,2 +1,0 @@
-# Used when this is a local dep, of e.g. test crates
-no_verify = true
