+++ prusti-contracts/prusti-contracts-proc-macros/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "prusti-contracts-internal"
+name = "prusti-contracts-proc-macros"
 version = "0.1.0"
 authors = ["Prusti Devs <prusti_developers@sympa.ethz.ch>"]
 edition = "2021"
