+++ prusti-contracts/prusti-contracts/Cargo.toml
@@ -12,11 +12,11 @@ keywords = ["prusti", "contracts", "verification", "formal", "specifications"]
 categories = ["development-tools", "development-tools::testing"]
 
 [dependencies]
-prusti-contracts-internal = { path = "../prusti-contracts-internal", version = "0.1.0" }
+prusti-contracts-proc-macros = { path = "../prusti-contracts-proc-macros", version = "0.1.0" }
 
 [dev-dependencies]
 trybuild = "1.0"
 
+# Forward "prusti" flag
 [features]
-# Are we being compiled by Prusti?
-prusti = ["prusti-contracts-internal/prusti"]
+prusti = ["prusti-contracts-proc-macros/prusti"]
