+++ prusti-contracts/prusti-std/Cargo.toml
@@ -12,8 +12,8 @@ keywords = ["prusti", "contracts", "verification", "formal", "specifications"]
 categories = ["development-tools", "development-tools::testing"]
 
 [dependencies]
-prusti-contracts = { path = "../prusti-contracts", version = "0.1.0" , features = ["prusti"] }
+prusti-contracts = { path = "../prusti-contracts", version = "0.1.0" }
 
+# Forward "prusti" flag
 [features]
-# Are we being compiled by Prusti?
 prusti = ["prusti-contracts/prusti"]
