+++ x.py
@@ -38,10 +38,9 @@ RUSTFMT_CRATES = [
     'prusti',
     #'prusti-common',
     'prusti-contracts/prusti-contracts',
-    'prusti-contracts/prusti-contracts-std',
-    'prusti-contracts/prusti-contracts-impl',
-    'prusti-contracts/prusti-contracts-internal',
+    'prusti-contracts/prusti-contracts-proc-macros',
     #'prusti-contracts/prusti-specs',
+    'prusti-contracts/prusti-std',
     'prusti-contracts-build',
     'prusti-interface',
     'prusti-launch',
